api: Add a reset callback
Add a reset callback that is called after the GPU's reset functions
have been called. This is called both when NoMali's API is used to
reset the GPU and when the GPU is reset using its register
interface. This functionality can, for example, be used to override
GPU ID registers using the nomali_reg_write_raw() interface.

Signed-off-by: Andreas Sandberg <andreas.sandberg@arm.com>